data_search_selenium/requirements.txt
selenium
requests
Pillow